Unify the r2lang logic into a separate singleton file
* Fixes double-r2spawn issue
